Added wfLYSO as a trait to GEEvents to hold the LYSO pre-amp traces. Created PairGEEvents binary to take two .root files with GEEvents and combine the two channels into one object. STRUCK parser now optionally takes the channel number as input so the channels can be parsed separatedly then paired afterwards.

+++ src/main/DoFloodCorrectionLYSO.cc
@@ -61,7 +61,7 @@ void DoFloodCorrection(string inFileName, string outFileName, string corrFileNam
 	for(int i=0;i < NUM_LYSO_PIXELS;i++)
 	{
 		getline(corrFile, buf);
-		double gc = (double)atoi(buf.c_str());
+		double gc = (double)atof(buf.c_str());
 		floodCorrect.push_back(gc);
 		cout << floodCorrect[i] << " ";
 	}
